versao final

+++ Inc/Load.h
@@ -6,6 +6,6 @@ void SET_CURRENT(float value);
 void SET_POTENCY(float potency);
 void SET_RESISTANCE(float resistance);
 void TURN_LOAD_OFF();
-int GET_CURRENT_SETED();
+float GET_CURRENT_SETED();
 
 #endif /* __LOAD_H */
